// src/types/instagram.d.ts
export type PostJson = {
  articleBody: string;
  author: Author;
  comment: Comment;
  commentCount: string;
  contentLocation: any;
  "@context": string;
  dateCreated: string;
  dateModified: string;
  headline: string;
  identifier: Identifier;
  image: any[];
  interactionStatistic: InteractionStatistic[];
  mainEntityOfPage: MainEntityOfPage;
  "@type": string;
  video: InstagramPageVideo[];
};

export type Author = {
  "@type": string;
  identifier: Identifier;
  image: string;
  name: string;
  alternateName: string;
  url: string;
};

export type Comment = {
  "@type": string;
  text: string;
  author: CommentAuthor;
  dateCreated: string;
  interactionStatistic: InteractionStatistic;
};

export type CommentAuthor = {
  "@type": string;
  identifier: Identifier;
  image: string;
  name: string;
  alternateName: string;
  url: string;
};

export type InteractionStatistic = {
  "@type": string;
  interactionType: string;
  userInteractionCount: number;
};

export type Identifier = {
  "@type": string;
  propertyID: string;
  value: string;
};

export type MainEntityOfPage = {
  "@type": string;
  "@id": string;
};

export type PostJsonVideo = {
  "@type": string;
  uploadDate: string;
  description: string;
  name: string;
  caption: string;
  height: string;
  width: string;
  contentUrl: string;
  thumbnailUrl: string;
  genre: string[];
  keywords: string[];
};

// GRAPHQL API TYPES

export interface GraphqlResponse {
  items: Item[];
  num_results: number;
  more_available: boolean;
  auto_load_more_enabled: boolean;
  showQRModal: boolean;
}

export interface GraphqlPostJson {
  taken_at: number;
  pk: number;
  id: string;
  device_timestamp: number;
  media_type: number;
  code: string;
  client_cache_key: string;
  filter_type: number;
  is_unified_video: boolean;
  should_request_ads: boolean;
  commerciality_status: string;
  is_visual_reply_commenter_notice_enabled: boolean;
  clips_tab_pinned_user_ids: any[];
  comment_inform_treatment: CommentInformTreatment;
  sharing_friction_info: SharingFrictionInfo;
  play_count: number;
  is_paid_partnership: boolean;
  has_delayed_metadata: boolean;
  comment_likes_enabled: boolean;
  comment_threading_enabled: boolean;
  max_num_visible_preview_comments: number;
  has_more_comments: boolean;
  preview_comments: any[];
  comments: any[];
  comment_count: number;
  can_view_more_preview_comments: boolean;
  hide_view_all_comment_entrypoint: boolean;
  inline_composer_display_condition: string;
  user: User;
  can_viewer_reshare: boolean;
  has_liked: boolean;
  like_count: number;
  top_likers: string[];
  facepile_top_likers: any[];
  image_versions2: ImageVersions2;
  original_width: number;
  original_height: number;
  is_organic_product_tagging_eligible: boolean;
  can_see_insights_as_brand: boolean;
  video_subtitles_confidence: number;
  video_subtitles_uri: string;
  caption: Caption;
  caption_is_edited: boolean;
  highlight_post_metadata: null;
  coauthor_producers: null;
  coauthor_producer_can_see_organic_insights: null;
  group: null;
  is_groups_post_pending_admin_approval: null;
  invited_coauthor_producers: null;
  guide_metadata: null;
  is_dash_eligible: number;
  video_dash_manifest: string;
  video_codec: string;
  number_of_qualities: number;
  video_versions: VideoVersion[];
  has_audio: boolean;
  video_duration: number;
  original_media_has_visual_reply_media: boolean;
  like_and_view_counts_disabled: boolean;
  can_viewer_save: boolean;
  is_in_profile_grid: boolean;
  profile_grid_control_enabled: boolean;
  featured_products: any[];
  is_comments_gif_composer_enabled: boolean;
  attribution: null;
  organic_tracking_token: string;
  has_shared_to_fb: number;
  product_type: string;
  show_shop_entrypoint: boolean;
  deleted_reason: number;
  integrity_review_decision: string;
  commerce_integrity_review_decision: null;
  music_metadata: null;
  is_artist_pick: boolean;
  ig_media_sharing_disabled: boolean;
  reshare_count: number;
  clips_metadata: ClipsMetadata;
  media_cropping_info: MediaCroppingInfo;
}

export interface Caption {
  pk: string;
  user_id: number;
  text: string;
  type: number;
  created_at: number;
  created_at_utc: number;
  content_type: string;
  status: string;
  bit_flags: number;
  did_report_as_spam: boolean;
  share_enabled: boolean;
  user: User;
  is_covered: boolean;
  is_ranked_comment: boolean;
  media_id: number;
  private_reply_status: number;
}

export interface User {
  has_anonymous_profile_picture: boolean;
  fan_club_info: FanClubInfo;
  fbid_v2: number;
  transparency_product_enabled: boolean;
  latest_reel_media: number;
  is_favorite: boolean;
  is_unpublished: boolean;
  pk: number;
  pk_id: string;
  strong_id__: string;
  username: string;
  full_name: string;
  is_private: boolean;
  is_verified: boolean;
  friendship_status: FriendshipStatus;
  profile_pic_id: string;
  profile_pic_url: string;
  account_badges: any[];
  show_account_transparency_details: boolean;
  third_party_downloads_enabled: number;
}

export interface FanClubInfo {
  fan_club_id: null;
  fan_club_name: null;
  is_fan_club_referral_eligible: null;
  fan_consideration_page_revamp_eligiblity: null;
  is_fan_club_gifting_eligible: null;
  subscriber_count: null;
}

export interface FriendshipStatus {
  following: boolean;
  outgoing_request: boolean;
  is_bestie: boolean;
  is_restricted: boolean;
  is_feed_favorite: boolean;
}

export interface ClipsMetadata {
  music_info: null;
  original_sound_info: OriginalSoundInfo;
  audio_type: string;
  music_canonical_id: string;
  featured_label: null;
  mashup_info: MashupInfo;
  reusable_text_info: null;
  reusable_text_attribute_string: null;
  nux_info: null;
  viewer_interaction_settings: null;
  branded_content_tag_info: BrandedContentTagInfo;
  shopping_info: null;
  additional_audio_info: AdditionalAudioInfo;
  is_shared_to_fb: boolean;
  breaking_content_info: null;
  challenge_info: null;
  reels_on_the_rise_info: null;
  breaking_creator_info: null;
  asset_recommendation_info: null;
  contextual_highlight_info: null;
  clips_creation_entry_point: string;
  audio_ranking_info: AudioRankingInfo;
  template_info: null;
  is_fan_club_promo_video: boolean;
  disable_use_in_clips_client_cache: boolean;
  content_appreciation_info: ContentAppreciationInfo;
  achievements_info: AchievementsInfo;
  show_achievements: boolean;
  show_tips: boolean;
  merchandising_pill_info: null;
  is_public_chat_welcome_video: boolean;
  professional_clips_upsell_type: number;
}

export interface AchievementsInfo {
  show_achievements: boolean;
  num_earned_achievements: null;
}

export interface AdditionalAudioInfo {
  additional_audio_username: null;
  audio_reattribution_info: AudioReattributionInfo;
}

export interface AudioReattributionInfo {
  should_allow_restore: boolean;
}

export interface AudioRankingInfo {
  best_audio_cluster_id: string;
}

export interface BrandedContentTagInfo {
  can_add_tag: boolean;
}

export interface ContentAppreciationInfo {
  enabled: boolean;
  entry_point_container: null;
}

export interface MashupInfo {
  mashups_allowed: boolean;
  can_toggle_mashups_allowed: boolean;
  has_been_mashed_up: boolean;
  formatted_mashups_count: null;
  original_media: null;
  privacy_filtered_mashups_media_count: null;
  non_privacy_filtered_mashups_media_count: number;
  mashup_type: null;
  is_creator_requesting_mashup: boolean;
  has_nonmimicable_additional_audio: boolean;
}

export interface OriginalSoundInfo {
  audio_asset_id: number;
  music_canonical_id: null;
  progressive_download_url: string;
  duration_in_ms: number;
  dash_manifest: string;
  ig_artist: IgArtist;
  should_mute_audio: boolean;
  hide_remixing: boolean;
  original_media_id: number;
  time_created: number;
  original_audio_title: string;
  consumption_info: ConsumptionInfo;
  can_remix_be_shared_to_fb: boolean;
  formatted_clips_media_count: null;
  allow_creator_to_rename: boolean;
  audio_parts: any[];
  is_explicit: boolean;
  original_audio_subtype: string;
  is_audio_automatically_attributed: boolean;
  is_reuse_disabled: boolean;
  is_xpost_from_fb: boolean;
  xpost_fb_creator_info: null;
  nft_info: null;
  is_original_audio_download_eligible: boolean;
}
// ...
export interface CommentInformTreatment {
  should_have_inform_treatment: boolean;
  text: string;
  url: null;
  action_type: null;
}

export interface ImageVersions2 {
  candidates: FirstFrame[];
  additional_candidates: AdditionalCandidates;
  smart_thumbnail_enabled: boolean;
  scrubber_spritesheet_info_candidates: ScrubberSpritesheetInfoCandidates;
}

export interface AdditionalCandidates {
  igtv_first_frame: FirstFrame;
  first_frame: FirstFrame;
  smart_frame: null;
}

export interface FirstFrame {
  width: number;
  height: number;
  url: string;
}

export interface ScrubberSpritesheetInfoCandidates {
  default: Default;
}

export interface Default {
  video_length: number;
  thumbnail_width: number;
  thumbnail_height: number;
  thumbnail_duration: number;
  sprite_urls: string[];
  thumbnails_per_row: number;
  total_thumbnail_num_per_sprite: number;
  max_thumbnails_per_sprite: number;
  sprite_width: number;
  sprite_height: number;
  rendered_width: number;
  file_size_kb: number;
}

export interface MediaCroppingInfo {
  feed_preview_crop: null;
  square_crop: SquareCrop;
  three_by_four_preview_crop: null;
}

export interface SquareCrop {
  crop_bottom: number;
  crop_left: number;
  crop_right: number;
  crop_top: number;
}

export interface SharingFrictionInfo {
  should_have_sharing_friction: boolean;
  bloks_app_url: null;
  sharing_friction_payload: null;
}

export interface VideoVersion {
  type: number;
  width: number;
  height: number;
  url: string;
  id: string;
}